LayoutTests/http/tests/websocket/tests/hixie76/websocket-protocol-ignored-expected.txt

Test WebSocket handshake success without protocol and ignore WebSocket-Protocol from server.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Connected
PASS protocol is "sub-protocol"
PASS successfullyParsed is true

TEST COMPLETE
